refactor(render): extract static pose bake into characters/static_pose_bake
Move bakeStaticPose and meshChainVisible out of characters/assets.ts into
their own sibling module so the far-LOD / shadow-proxy bake is unit-testable
without the loader and preload machinery. Verbatim move except two comment
dashes rewritten as ASCII per the repo copy rule. No behavior change.
