+++ src/render/characters/assets.ts
@@ -8,7 +8,6 @@
 // so everything here can assume resolved GLTFs synchronously afterwards.
 import * as THREE from 'three';
 import type { GLTF } from 'three/addons/loaders/GLTFLoader.js';
-import { mergeGeometries } from 'three/addons/utils/BufferGeometryUtils.js';
 import { clone as cloneSkinned } from 'three/addons/utils/SkeletonUtils.js';
 import { WEAPON_SKINS } from '../../sim/content/weapon_skins';
 import { loadGltf, loadTexture } from '../assets/loader';
@@ -28,6 +27,7 @@ import {
 } from './manifest';
 import { mergeSkinnedParts } from './rig_merge';
 import { weaponSkinAttachBone, weaponSkinHandling } from './skin_attack';
+import { bakeStaticPose, meshChainVisible } from './static_pose_bake';
 import { variantGripTransform, WEAPON_GRIP_OVERRIDES } from './weapon_grip';
 import { markOwnedWeaponSkinMaterials } from './weapon_skin_materials';
 
@@ -907,69 +907,3 @@ export function prepareVisual(key: string): PreparedVisual {
   return prep;
 }
 
-function meshChainVisible(o: THREE.Object3D, stopAt: THREE.Object3D): boolean {
-  let cur: THREE.Object3D | null = o;
-  while (cur) {
-    if (!cur.visible) return false;
-    if (cur === stopAt) return true;
-    cur = cur.parent;
-  }
-  return true;
-}
-
-/** Bake every visible mesh of a posed clone into one static BufferGeometry
- *  (skinned verts via applyBoneTransform), normalized into world units. */
-function bakeStaticPose(
-  root: THREE.Object3D,
-  norm: THREE.Matrix4,
-): { geo: THREE.BufferGeometry | null; mats: THREE.Material[] } {
-  const geos: THREE.BufferGeometry[] = [];
-  const mats: THREE.Material[] = [];
-  const v = new THREE.Vector3();
-  const full = new THREE.Matrix4();
-
-  root.traverse((o) => {
-    const mesh = o as THREE.Mesh;
-    if (!mesh.isMesh || !meshChainVisible(mesh, root)) return;
-    const srcGeo = mesh.geometry;
-    const srcPos = srcGeo.getAttribute('position') as THREE.BufferAttribute;
-    if (!srcPos) return;
-    const out = new THREE.BufferGeometry();
-    const baked = new Float32Array(srcPos.count * 3);
-    const skinned = (mesh as unknown as THREE.SkinnedMesh).isSkinnedMesh
-      ? (mesh as unknown as THREE.SkinnedMesh)
-      : null;
-    full.multiplyMatrices(norm, mesh.matrixWorld);
-    for (let i = 0; i < srcPos.count; i++) {
-      v.fromBufferAttribute(srcPos, i);
-      if (skinned) {
-        skinned.applyBoneTransform(i, v);
-        v.applyMatrix4(skinned.matrixWorld).applyMatrix4(norm);
-      } else {
-        v.applyMatrix4(full);
-      }
-      baked[i * 3] = v.x;
-      baked[i * 3 + 1] = v.y;
-      baked[i * 3 + 2] = v.z;
-    }
-    out.setAttribute('position', new THREE.BufferAttribute(baked, 3));
-    const uv = srcGeo.getAttribute('uv');
-    if (uv) out.setAttribute('uv', uv.clone());
-    if (srcGeo.index) out.setIndex(srcGeo.index.clone());
-    out.computeVertexNormals();
-    geos.push(out);
-    // GLTFLoader emits one Mesh per primitive — materials are never arrays here
-    mats.push(Array.isArray(mesh.material) ? mesh.material[0] : mesh.material);
-  });
-
-  if (geos.length === 0) return { geo: null, mats: [] };
-  // uv presence must agree for merging — drop uvs entirely if any geo lacks them
-  const allHaveUv = geos.every((g) => g.getAttribute('uv'));
-  if (!allHaveUv) for (const g of geos) g.deleteAttribute('uv');
-  const geo = geos.length === 1 ? geos[0] : mergeGeometries(geos, true);
-  if (geos.length === 1) {
-    geo.clearGroups();
-    geo.addGroup(0, geo.index ? geo.index.count : geo.getAttribute('position').count, 0);
-  }
-  return { geo, mats };
-}
